CI: optimize parallel build steps and add demo integration testing

+++ hack/lib/artifacts.sh
@@ -5,8 +5,12 @@ set -o nounset
 set -o pipefail
 
 ARTIFACTS=${REPO_ROOT}/artifacts
+ARTIFACTS_CLEAR_MODE="clear"
+
 mkdir -p ${ARTIFACTS}
 
 trisa::artifacts::clear() {
-   rm -rf ${ARTIFACTS}/*
+   if [ "${ARTIFACTS_CLEAR_MODE}" == "clear" ]; then
+      rm -rf ${ARTIFACTS}/*
+   fi
 }
